Also escape `(`, `)` and `.` in browser filenames
Got the following error:

```
21 07 2015 10:11:09.382:WARN [reporter.junit]: Cannot write JUnit xml
	ENOENT, open '/dev/gulp/tmp/PhantomJS_1.9.8_(Mac_OS_X_0.0.0)/test_result.xml'
```

Tried creating the directory manually, and got the following error:

```bash
$ mkdir PhantomJS_1.9.8_(Mac_OS_X_0.0.0)
bash: syntax error near unexpected token `('
```

So, I'd guessed that we needed to escape those characters too. A little escaping doesn't hurt now, does it?

I'm not sure if there are more troublesome characters, and if we therefor want to change the mkdir implementation to better escape, but for now I'm just focussed on getting this working for me. Let me know what you think!